refactor(metadata): migrate inventorychecks component to v2
