--- comp/metadata/inventorychecks/component_mock.go
@@ -1,11 +1,0 @@
-// Unless explicitly stated otherwise all files in this repository are licensed
-// under the Apache License Version 2.0.
-// This product includes software developed at Datadog (https://www.datadoghq.com/).
-// Copyright 2016-present Datadog, Inc.
-
-package inventorychecks
-
-// Mock implements mock-specific methods for the inventorychecks component.
-type Mock interface {
-	Component
-}
