+++ comp/metadata/inventorychecks/def/component.go
