+++ comp/metadata/inventorychecks/impl/integrations_nojmx.go
@@ -5,7 +5,7 @@
 
 //go:build !jmx
 
-package inventorychecksimpl
+package inventorycheckimpl
 
 func (ic *inventorychecksImpl) getJMXChecksMetadata() (jmxMetadata map[string][]metadata) {
 	// This function is a no-op when JMX is not enabled.
